requirements.txt
fastapi
uvicorn
requests
python-dotenv
psycopg2-binary
sentence-transformers
qdrant-client
numpy
unicodedata2
beautifulsoup4
concurrent-log-handler
pandas
accelerate
langchain-core
langchain-community
langgraph
langchain-huggingface
langchain-community
onnx
optimum[onnxruntime]
tensorrt
onnx-graphsurgeon
openpyxl
locust
vllm
duckduckgo-search
uv
fitz
fronend
pdfplumber
langsmith 
openevals
nltk
google-generativeai